Rough IDA 9.0 compat
I'm sprinkling int() casts everywhere, probably way too many,
but I'm too lazy to figure out exactly where it's unnecessary! =)

+++ plugins/tenet/ui/reg_view.py
@@ -20,7 +20,7 @@ class RegisterView(QtWidgets.QWidget):
         # child widgets
         self.reg_area = RegisterArea(self.controller, self.model, self)
         self.idx_shell = TimestampShell(self.controller, self.model, self)
-        self.setMinimumWidth(self.reg_area.minimumWidth())
+        self.setMinimumWidth(int(self.reg_area.minimumWidth()))
 
         # layout
         layout = QtWidgets.QVBoxLayout(self)
@@ -100,7 +100,7 @@ class RegisterArea(QtWidgets.QAbstractScrollArea):
         self._char_height = fm.height()
 
         # default to fit roughly 50 printable characters
-        self._default_width = self._char_width * (self.pctx.arch.POINTER_SIZE * 2 + 16)
+        self._default_width = int(self._char_width * (self.pctx.arch.POINTER_SIZE * 2 + 16))
 
         # register drawing information
         self._reg_pos = (self._char_width, self._char_height)
@@ -109,7 +109,7 @@ class RegisterArea(QtWidgets.QAbstractScrollArea):
 
         self.setFocusPolicy(QtCore.Qt.StrongFocus)
         self.setVerticalScrollBarPolicy(QtCore.Qt.ScrollBarAsNeeded)
-        self.setMinimumWidth(self._reg_pos[0] + self._default_width)
+        self.setMinimumWidth(int(self._reg_pos[0] + self._default_width))
         self.setMouseTracking(True)
 
         self._init_ctx_menu()
@@ -118,8 +118,8 @@ class RegisterArea(QtWidgets.QAbstractScrollArea):
         self.model.registers_changed(self.refresh)
 
     def sizeHint(self):
-        width = self._default_width
-        height = (len(self._reg_fields) + 2) * self._char_height # +2 for line break before IP, and after IP
+        width = int(self._default_width)
+        height = int((len(self._reg_fields) + 2) * self._char_height) # +2 for line break before IP, and after IP
         return QtCore.QSize(width, height)
 
     def _init_ctx_menu(self):
@@ -162,22 +162,22 @@ class RegisterArea(QtWidgets.QAbstractScrollArea):
             if reg_name == self.model.arch.IP:
                 y += self._char_height
 
-            name_rect = QtCore.QRect(0, 0, name_size.width(), name_size.height())
-            name_rect.moveBottomLeft(QtCore.QPoint(name_x, y))
+            name_rect = QtCore.QRect(0, 0, int(name_size.width()), int(name_size.height()))
+            name_rect.moveBottomLeft(QtCore.QPoint(int(name_x), int(y)))
 
             prev_rect = QtCore.QRect(0, 0, arrow_size, arrow_size)
             next_rect = QtCore.QRect(0, 0, arrow_size, arrow_size)
             arrow_rects = [prev_rect, next_rect]
 
-            prev_x = name_x + name_size.width() + self._char_width
+            prev_x = int(name_x + name_size.width() + self._char_width)
             prev_rect.moveCenter(name_rect.center())
             prev_rect.moveLeft(prev_x)
 
             value_x = prev_x + prev_rect.width() + self._char_width
-            value_rect = QtCore.QRect(0, 0, value_size.width(), value_size.height())
-            value_rect.moveBottomLeft(QtCore.QPoint(value_x, y))
+            value_rect = QtCore.QRect(0, 0, int(value_size.width()), int(value_size.height()))
+            value_rect.moveBottomLeft(QtCore.QPoint(int(value_x), int(y)))
 
-            next_x = value_x + value_size.width() + self._char_width
+            next_x = int(value_x + value_size.width() + self._char_width)
             next_rect.moveCenter(name_rect.center())
             next_rect.moveLeft(next_x)
 
@@ -258,7 +258,7 @@ class RegisterArea(QtWidgets.QAbstractScrollArea):
         Get the register field at the given cursor position.
         """
         for reg_name, field in self._reg_fields.items():
-            full_field = QtCore.QRect(field.name_rect.topLeft(), field.next_rect.bottomRight())
+            full_field = QtCore.QRect(int(field.name_rect.topLeft()), int(field.next_rect.bottomRight()))
             if full_field.contains(pos):
                 return field
         return None
@@ -274,8 +274,8 @@ class RegisterArea(QtWidgets.QAbstractScrollArea):
         if not self.model.registers:
             return QtCore.QSize(0, 0)
 
-        width = self._reg_pos[0] + self._default_width
-        height = len(self.model.registers) * self._char_height
+        width = int(self._reg_pos[0] + self._default_width)
+        height = int(len(self.model.registers) * self._char_height)
 
         return QtCore.QSize(width, height)
